+++ db/migrate/20200709161504_users.rb
@@ -1,0 +1,7 @@
+class Users< ActiveRecord::Migration[5.2]
+  def change
+    create_table :users do |c|
+    c.string :name
+    end
+  end
+end
