+++ lib/film.rb
@@ -42,7 +42,7 @@ class Film < ActiveRecord::Base
     end
 
     # Person.all.delete(Person.find_by(name: "Ashitaka"))
-    # Film.find_by(title: ).description
+
 
     # - May make sense to skip this one. API only provides a link to locations, but no identifying information.
     # - We'd have to go over the films array and match the original film ID to the end of the string,
