+++ lib/person.rb
@@ -1,16 +1,12 @@
 class Person < ActiveRecord::Base
     belongs_to :type
     belongs_to :film
+    belongs_to :user
 
     def self.get_people
         people = RestClient.get('https://ghibliapi.herokuapp.com/people')
         people = JSON.parse(people)
         people
     end    
-    
-    # eye_color
-    # hair_color
-    # gender
-    # age
 end
