+++ lib/user.rb
@@ -1,0 +1,41 @@
+class User < ActiveRecord::Base
+    has_many :people
+    has_many :films, through: :people
+
+    def finder
+        my_world_info = Person.all.where(user_id: 1)
+    end
+
+    def get_people_by_user
+        finder.map do |info|
+            info.name
+        end
+    end 
+
+    def get_films_by_user
+        film_titles = []
+        finder.each do |info|
+           film_titles << Film.find_by(id: info.film_id).title
+        end
+        film_titles.uniq
+    end
+
+    def get_types_by_user
+        type_names = []
+        finder.each do |info|
+            type_names << Type.find_by(id: info.type_id).name
+        end
+        type_names.uniq
+    end
+
+# Person.all.update(user_id: 1)
+    def starter_personal_data
+        my_characters = []
+        Person.all.each do |info|
+            my_characters << info.dup.update(user_id: 1)
+        end
+        my_characters
+    end
+
+
+end
